requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0 